+++ src/electrifier/Controls/Vanara/ExplorerBrowser.xaml.cs
@@ -14,6 +14,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices.WindowsRuntime;
 using System.Runtime.InteropServices;
 using System.Windows.Input;
+using electrifier.Controls.Vanara.Services;
 namespace electrifier.Controls.Vanara;
 using Visibility = Microsoft.UI.Xaml.Visibility;
 // https://github.com/dahall/Vanara/blob/master/Windows.Forms/Controls/ExplorerBrowser.cs
@@ -199,8 +200,6 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
     {
         get;
     }
-    private static SoftwareBitmapSource? _defaultFolderImageBitmapSource;
-    private static SoftwareBitmapSource? _defaultDocumentAssocImageBitmapSource;
     /// <summary>Raises the <see cref="NavigationFailed"/> event.</summary>
     internal void OnNavigationFailed(ExtNavigationFailedEventArgs? nfevent)
     {
@@ -213,6 +212,7 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
     }
     /// <summary>Fires when either a Navigating listener cancels the navigation, or if the operating system determines that navigation is not possible.</summary>
     public event EventHandler<ExtNavigationFailedEventArgs>? NavigationFailed;
+    private ShellNamespaceService _nsService = new ShellNamespaceService();
 
     /// <summary>
     /// ExplorerBrowser Implementation for WinUI3
@@ -257,8 +257,6 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
 
     private async Task InitializeViewModel()
     {
-        _stockIconTask = InitializeStockIcons();
-
         var rootItems = new List<ExplorerBrowserItem>
         {
             // todo: new ExplorerBrowserItem(HomeShellFolder.PIDL),
@@ -279,36 +277,6 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
         ShellTreeView.ItemsSource = rootItems;
     }
 
-    /// <summary>
-    /// <see href="https://github.com/dahall/Vanara/blob/ac0a1ac301dd4fdea9706688dedf96d596a4908a/Windows.Shell.Common/StockIcon.cs"/>
-    /// </summary>
-    /// <returns></returns>
-    public async Task InitializeStockIcons()
-    {
-        try
-        {
-            using var siFolder = new StockIcon(Shell32.SHSTOCKICONID.SIID_FOLDER);
-            {
-                var idx = siFolder.SystemImageIndex;
-                var icnHandle = siFolder.IconHandle.ToIcon();
-                var bmpSource = GetWinUi3BitmapSourceFromIcon(icnHandle);
-                _defaultFolderImageBitmapSource = await bmpSource;
-            }
-
-            using var siDocument = new StockIcon(Shell32.SHSTOCKICONID.SIID_DOCNOASSOC);
-            {
-                var idx = siDocument.SystemImageIndex;
-                var icnHandle = siDocument.IconHandle.ToIcon();
-                var bmpSource = GetWinUi3BitmapSourceFromIcon(icnHandle);
-                _defaultDocumentAssocImageBitmapSource = await bmpSource;
-            }
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e);
-            throw;
-        }
-    }
     private void ExplorerBrowser_NavigationFailed(object? sender, ExtNavigationFailedEventArgs e)
     {
         var location = e.FailedLocation;
@@ -362,12 +330,10 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
                     var ebItem = new ExplorerBrowserItem(item.PIDL);
                     if (ebItem.IsFolder)
                     {
-                        ebItem.BitmapSource = _defaultFolderImageBitmapSource;
                         result.Insert(0, ebItem);
                     }
                     else
                     {
-                        ebItem.BitmapSource = _defaultDocumentAssocImageBitmapSource;
                         result.Add(ebItem);
                     }
                 }
@@ -502,50 +468,6 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
             IsLoading = false;
 
         }
-    }
-
-    /// <summary>
-    /// Taken from <see href="https://stackoverflow.com/questions/76640972/convert-system-drawing-icon-to-microsoft-ui-xaml-imagesource"/>
-    /// See also <see href="https://learn.microsoft.com/en-us/windows/windows-app-sdk/api/winrt/microsoft.ui.xaml.media.imaging.bitmapimage?view=windows-app-sdk-1.6"/>, which can deal with .ico natively.
-    /// </summary>
-    /// <param name="bitmapIcon"></param>
-    /// <returns></returns>
-    public static async Task<SoftwareBitmapSource?> GetWinUi3BitmapSourceFromIcon(Icon bitmapIcon)
-    {
-        ArgumentNullException.ThrowIfNull(bitmapIcon);
-
-        return await GetWinUi3BitmapSourceFromGdiBitmap(bitmapIcon.ToBitmap());
-    }
-
-    /// <summary>
-    /// Taken from <see href="https://stackoverflow.com/questions/76640972/convert-system-drawing-icon-to-microsoft-ui-xaml-imagesource"/>.
-    /// See also <see href="https://learn.microsoft.com/en-us/windows/windows-app-sdk/api/winrt/microsoft.ui.xaml.controls.image.source?view=windows-app-sdk-1.5#microsoft-ui-xaml-controls-image-source"/>.
-    /// See also <see href="https://learn.microsoft.com/en-us/windows/windows-app-sdk/api/winrt/microsoft.ui.xaml.media.imaging.bitmapimage?view=windows-app-sdk-1.6"/>, which can deal with .ico natively.
-    /// </summary>
-    /// <param name="gdiBitmap"></param>
-    /// <returns></returns>
-    public static async Task<SoftwareBitmapSource?> GetWinUi3BitmapSourceFromGdiBitmap(Bitmap gdiBitmap)
-    {
-        ArgumentNullException.ThrowIfNull(gdiBitmap);
-
-        // get pixels as an array of bytes
-        var data = gdiBitmap.LockBits(new Rectangle(0, 0, gdiBitmap.Width, gdiBitmap.Height), System.Drawing.Imaging.ImageLockMode.ReadOnly, gdiBitmap.PixelFormat);
-        var bytes = new byte[data.Stride * data.Height];
-        Marshal.Copy(data.Scan0, bytes, 0, bytes.Length);
-        gdiBitmap.UnlockBits(data);
-
-        // get WinRT SoftwareBitmap
-        var softwareBitmap = new Windows.Graphics.Imaging.SoftwareBitmap(
-            Windows.Graphics.Imaging.BitmapPixelFormat.Bgra8,
-            gdiBitmap.Width,
-            gdiBitmap.Height,
-            Windows.Graphics.Imaging.BitmapAlphaMode.Premultiplied);
-        softwareBitmap.CopyFromBuffer(bytes.AsBuffer());
-
-        // build WinUI3 SoftwareBitmapSource
-        var source = new SoftwareBitmapSource();
-        await source.SetBitmapAsync(softwareBitmap);
-        return source;
     }
 
     public void OnRefreshViewCommand(object sender, RoutedEventArgs e)
